Added Mid Shield Cracking
A bit comical, but we can change the hitbox detection. Also edited the
materials on the different parts of the reactor and moved a script
around.
