+++ Assets/Materials/Reactor.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 33af5f4726fe6994898cd0d7a93fa861
+folderAsset: yes
+timeCreated: 1461294614
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
