+++ Assets/Materials/Reactor/Shield2.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5d1728585754ab444bb5ace81fe14645
+timeCreated: 1461291318
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
