--- Assets/Models/ReactorBoss/Materials/Shield1.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bf08af5ffe29a9148acfa8b656776eda
-timeCreated: 1461291318
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
