--- Assets/Models/ReactorBoss/Materials/ShieldOuter.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a477b22fe7ba89c4694f3a91858c475f
-timeCreated: 1461291317
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
