--- Assets/ShatterWhenHit.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 1265d541e0660ff4ab86ef5b4308776e
-timeCreated: 1461292559
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
